+++ src/components/landing/AdvantagesInfo.tsx
