+++ src/components/landing/GettingStarted.tsx
